+++ backend/utils/log.py
@@ -3,7 +3,7 @@ import sys
 import logging
 from loguru import logger
 
-from backend.config.config import LOG_LEVEL, LOG_TO_CONSOLE, LOG_TO_FILE, LOG_FILE, LOG_FORMAT
+from config.config import LOG_LEVEL, LOG_TO_CONSOLE, LOG_TO_FILE, LOG_FILE, LOG_FORMAT
 
 log_file = LOG_FILE
